+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/buttons/InlineKeyboardMarkup.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.buttons
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardButtons.InlineKeyboardButton
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardButtons.PayInlineKeyboardButton
 import com.github.insanusmokrassar.TelegramBotAPI.types.inlineKeyboardField
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -9,4 +10,19 @@ import kotlinx.serialization.Serializable
 data class InlineKeyboardMarkup(
     @SerialName(inlineKeyboardField)
     val keyboard: Matrix<InlineKeyboardButton>
-) : KeyboardMarkup
+) : KeyboardMarkup {
+    init {
+        val isTherePayButton = keyboard.any { it ->
+            it.any {
+                it is PayInlineKeyboardButton
+            }
+        }
+        if (isTherePayButton) {
+            // first button is not PayInlineKeyboardButton
+            val firstIsPaymentButton = keyboard.first().firstOrNull() is PayInlineKeyboardButton
+            if (!firstIsPaymentButton) {
+                error("In case if PayInlineKeyboardButton included in keyboard - it must ")
+            }
+        }
+    }
+}