+++ TelegramBotAPI/src/jvmMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/abstracts/InputFileFromJavaFile.kt
@@ -3,6 +3,10 @@ package com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts
 import com.github.insanusmokrassar.TelegramBotAPI.utils.StorageFile
 import java.io.File
 
-fun File.toInputFile() = MultipartFile(
-    StorageFile(this)
-)
+fun File.toInputFile() = if (exists()) {
+    MultipartFile(
+        StorageFile(this)
+    )
+} else {
+    error("Specified file $absolutePath does not exists")
+}